vla/src/vision-pipeline/requirements-runpod.txt
# Minimal deps for serverless inference (no WebRTC build requirements)
# Serverless + WebRTC + gRPC runtime
runpod==1.3.1
numpy>=1.24,<2.0
Pillow>=9.0,<11.0
onnxruntime-gpu>=1.17,<1.19
opencv-python>=4.8,<4.10
aiohttp>=3.8,<3.9
aiortc>=1.5,<2.0
av>=10.0,<12.0
grpcio>=1.60,<1.63
protobuf>=4.23,<5.0
